To Keep GitHub action up-to-date with GitHub Dependabot
If you opted for the GitHub action method, you might consider using Dependabot for this repo. It will make sure you are always using the latest version of our GitHub action. ‍ To enable Dependabot for this repository, all you need to do is add a .github/dependabot.yml file with the following contents.